RFC: promo up
Hello @nerves-project!

A long time ago, Elixir had a home page quite similar to Nerves. We focused a lot on the implementation details and how things worked instead of focusing on what people can effectively build with it. For example, we said Elixir was a dynamic language with pattern matching and protocols. We eventually improved it based on external feedback.

Therefore the goal with this pull request is to apply the same feedback Elixir got in the past to Nerves. I.e. focus a bit more on what people can achieve with Nerves and, even more importantly, all the things they *do not need to do* because Nerves solves it for them.

Those are just ideas. There are probably innacuracies too, so feel free to adapt or fully discard them. Btw, it would also be great to include a link to the "rock-solid shipping industrial products". :)